Querying sql in python
More practices. In this file, I created a database, created a table, added a variable that I filled with random numbers using a python module numpy random. I copied a subset of the observations from that database to another database. I also created a set of random number to a text file. Then copied the information from that file into a sql database... So basically, I was making use of sqlite3 and running sql from python.